% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{get_cran_deps}
\alias{get_cran_deps}
\title{Recursive (hard) dependencies of CRAN packages}
\usage{
get_cran_deps(packages)
}
\arguments{
\item{packages}{Character vector of package names.}
}
\value{
A character vector of package names and versions,
  separated by a single dash.
}
\description{
Recursive (hard) dependencies of CRAN packages
}
\keyword{internal}
