% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysreqs.R
\name{get_cran_sysreqs}
\alias{get_cran_sysreqs}
\title{Get the system requirements of a CRAN package}
\usage{
get_cran_sysreqs(package, platform = current_platform())
}
\arguments{
\item{package}{Package name.}

\item{platform}{Platform string. Defaults to the current platform.}
}
\value{
All system dependencies on the current or the specified
  platform.
}
\description{
Get the system requirements of a CRAN package
}
\keyword{internal}
