% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{get_remote_deps}
\alias{get_remote_deps}
\title{Recursive (hard) dependencies of remote packages}
\usage{
get_remote_deps(remote_packages)
}
\arguments{
\item{remote_packages}{Character vector of package names.}
}
\value{
A character vector of package names and versions,
  separated by a single dash.
}
\description{
Recursive (hard) dependencies of remote packages
}
\keyword{internal}
