% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{sysreq_commands}
\alias{sysreq_commands}
\title{Command(s) to install system requirements of a package, on a platform}
\usage{
sysreq_commands(desc, platform = current_platform(), soft = TRUE)
}
\arguments{
\item{desc}{Path to a \code{DESCRIPTION} file.}

\item{platform}{Platform string, defaults to the current platform.}

\item{soft}{Whether to include soft dependencies.}
}
\value{
Character scalar, a shell command to run to install
  the system requirements on a given platform.
}
\description{
Command(s) to install system requirements of a package, on a platform
}

